refactor(bytecode): 修正类型定义
- 将 NEG_DOUBLE、NEG_FLOAT、ATOMIC_NEG_DOUBLE 和 ATOMIC_NEG_FLOAT 的类型从 uint64_t 改为 uint8_t

+++ bytecode.h
@@ -169,10 +169,10 @@ namespace lvm::bytecode
     constexpr uint8_t JUMP_IF_FALSE = 0x82;
     constexpr uint8_t SYSCALL = 0x83;
     constexpr uint8_t THREAD_FINISH = 0x84;
-    constexpr uint64_t NEG_DOUBLE = 0x85;
-    constexpr uint64_t NEG_FLOAT = 0x86;
-    constexpr uint64_t ATOMIC_NEG_DOUBLE = 0x87;
-    constexpr uint64_t ATOMIC_NEG_FLOAT = 0x88;
+    constexpr uint8_t NEG_DOUBLE = 0x85;
+    constexpr uint8_t NEG_FLOAT = 0x86;
+    constexpr uint8_t ATOMIC_NEG_DOUBLE = 0x87;
+    constexpr uint8_t ATOMIC_NEG_FLOAT = 0x88;
 
     std::string getInstructionName(uint8_t code);
     uint8_t parseInstructionCode(std::string code);
